+++ src/com/sekwah/advancedportals/destinations/WarpCommand.java
@@ -1,6 +1,8 @@
-package com.sekwah.advancedportals;
+package com.sekwah.advancedportals.destinations;
 
-import com.sekwah.advancedportals.destinations.Destination;
+import com.sekwah.advancedportals.AdvancedPortalsPlugin;
+import com.sekwah.advancedportals.ConfigAccessor;
+import com.sekwah.advancedportals.PluginMessages;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -29,26 +31,9 @@ public class WarpCommand implements CommandExecutor, TabCompleter {
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String command, String[] args) {
         if (args.length > 0) {
-            if (Destination.warp(sender, args[0], false)) {
-                if (DestinationCommand.PortalMessagesDisplay == 1) {
-                    sender.sendMessage("");
-                    sender.sendMessage(plugin.customPrefixFail + "\u00A7a You have been warped to \u00A7e" + args[0].replaceAll("_", " ") + "\u00A7a.");
-                    sender.sendMessage("");
-                } else if (DestinationCommand.PortalMessagesDisplay == 2) {
-                    ConfigAccessor config = new ConfigAccessor(plugin, "destinations.yml");
-                    plugin.nmsAccess.sendActionBarMessage("{\"text\":\"\u00A7aYou have warped to \u00A7e" + args[0].replaceAll("_", " ") + "\u00A7a.\"}", (Player) sender);
-                    /**plugin.nmsAccess.sendActionBarMessage("[{text:\"You have warped to \",color:green},{text:\"" + config.getConfig().getString(Portal.Portals[portalId].portalName + ".destination").replaceAll("_", " ")
-                     + "\",color:yellow},{\"text\":\".\",color:green}]", player);*/
-                }
-            } else {
-                sender.sendMessage("");
-                sender.sendMessage(plugin.customPrefixFail + "\u00A7c The destination you tried to warp to does not exist!");
-                sender.sendMessage("");
-            }
+            Destination.warp((Player) sender, args[0]);
         } else {
-            sender.sendMessage("");
-            sender.sendMessage(plugin.customPrefixFail + "\u00A7c You need to type a destination after /" + command + "!");
-            sender.sendMessage("");
+            PluginMessages.UnknownCommand(sender, command);
         }
         return true;
     }
